[#noissue] Extract ServerTime module
