+++ web/src/main/java/com/navercorp/pinpoint/web/servertime/ServerTimeConfiguration.java
@@ -1,11 +1,11 @@
 /*
- * Copyright 2014 NAVER Corp.
+ * Copyright 2025 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,23 +14,13 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.web.view;
+package com.navercorp.pinpoint.web.servertime;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
 
-/**
- * @author emeroad
- */
-public class ServerTime {
+@ComponentScan(basePackages = "com.navercorp.pinpoint.web.controller")
+@Configuration
+public class ServerTimeConfiguration {
 
-    private final long currentServerTime;
-
-    public ServerTime() {
-        this.currentServerTime = System.currentTimeMillis();
-    }
-
-    @JsonProperty("currentServerTime")
-    public long getCurrentServerTime() {
-        return currentServerTime;
-    }
 }
