+++ web/src/main/java/com/navercorp/pinpoint/web/servertime/controller/ServerTimeController.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2025 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.web.servertime.controller;
+
+import com.navercorp.pinpoint.web.servertime.view.ServerTime;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@Validated
+public class ServerTimeController {
+
+    @GetMapping(value = {"/api/serverTime", "/api-public/serverTime"})
+    public ServerTime getServerTime() {
+        return new ServerTime();
+    }
+
+}
