+++ web/src/main/java/com/navercorp/pinpoint/web/servertime/view/ServerTime.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2025 NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.pinpoint.web.servertime.view;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * @author emeroad
+ */
+public class ServerTime {
+
+    private final long currentServerTime;
+
+    public ServerTime() {
+        this.currentServerTime = System.currentTimeMillis();
+    }
+
+    @JsonProperty("currentServerTime")
+    public long getCurrentServerTime() {
+        return currentServerTime;
+    }
+}
